refactor: 전체 폴더 구조 변경
