--- src/entities/product/response.ts
@@ -1,6 +1,0 @@
-import {
-  BaseProduct,
-  PaginatedResponse,
-} from "../../shared/api/types/response";
-
-export type Products = PaginatedResponse<BaseProduct>;
