+++ src/shared/hooks/useModal.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ModalContext } from "../../app/providers/ModalContext";
+import { ModalContext } from "../contexts/ModalContext";
 
 const useModal = () => {
   const context = useContext(ModalContext);
